Move tests to proper files
